Message resolver improvements
Hierarchical message property matching:

e.g. if a field exists:

a.b.Contact.firstName

then if a messages.properties entry exists with the key:

a.b.Contact.firstName

it will be used, but then matches will be looked for, starting from the end, e.g. the first match will be used of the following:

firstName
Contact.firstName
b.Contact.firstName

If still no match, then a default is used. For field labels that's the field name, for placeholders that's an empty string.

This allows flexible reuse of messages and allowing overrides where needed.