@Select can take a set of options
Boolean doesn't currently work without a @Select annotation.

On the plus side, you can do this in your form:

    @Select({ "true", "false" })
    private Boolean inCurrentUse;

or

    @Select({"red", "green", "blue", "unknown"})
    private String something;

And the parameter to @Select will be used for *values* in the dropdown's options.

Provided the values are valid property keys, then you can resolve them from, e.g. messages.properties:

    inCurrentUse=In Current Use?
    inCurrentUse.labels.true=Actively being used
    inCurrentUse.labels.false=In storage

or:

    something=Something
    something.labels.red=Burgundy
    something.labels.green=Green
    something.labels.blue=Navy